Migrate remaining regression tests
